+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - `Innmind\Http\Header::values()` now returns an `Innmind\Immutable\Sequence` to allow for ordered values
 - `Innmind\Http\Header\AcceptRanges` constructor is now private, use `::of()` or `::maybe()` named constructors
 - `Innmind\Http\Header\Age` constructor is now private, use `::of()` or `::maybe()` named constructors
+- `Innmind\Http\Header\Authorization` constructor is now private, use `::of()` or `::maybe()` named constructors
 
 ### Removed
 
@@ -32,6 +33,7 @@
 - `Innmind\Http\Factory\HeaderFactory`
 - `Innmind\Http\Header\AcceptRangesValue`
 - `Innmind\Http\Header\AgeValue`
+- `Innmind\Http\Header\AuthorizationValue`
 
 ### Fixed